fix player_index that was int and should have been size_t

+++ executables/LUDUMDARE/Ludum43/src/Ludum43Game.cpp
@@ -338,7 +338,7 @@ void LudumGame::OnLevelChanged(death::Level * new_level, death::Level * old_leve
         PlaySound("next_level", false, false, 0.0f, death::SoundContext::GAME);
 }
 
-float LudumGame::GetPlayerParticleLife(int player_index) const
+float LudumGame::GetPlayerParticleLife(size_t player_index) const
 {
 	ParticlePlayer const * player_particle = GetPlayerParticle(player_index);
 	if (player_particle == nullptr)
@@ -346,7 +346,7 @@ float LudumGame::GetPlayerParticleLife(int player_index) const
 	return player_particle->life;
 }
 
-ParticlePlayer * LudumGame::GetPlayerParticle(int player_index)
+ParticlePlayer * LudumGame::GetPlayerParticle(size_t player_index)
 {
 	LudumPlayer * player = GetLudumPlayer(player_index);
 	if (player == nullptr)
@@ -354,7 +354,7 @@ ParticlePlayer * LudumGame::GetPlayerParticle(int player_index)
 	return player->GetPlayerParticle();
 }
 
-ParticlePlayer const * LudumGame::GetPlayerParticle(int player_index) const
+ParticlePlayer const * LudumGame::GetPlayerParticle(size_t player_index) const
 {
 	LudumPlayer const * player = GetLudumPlayer(player_index);
 	if (player == nullptr)
